+++ NEWS.md
@@ -3,6 +3,7 @@
 * Fix order_type() for v4 (#18)
 * Tests
 * Multi-account batching for v3 (#17)
+* Add BigQuery fetching for GoogleAnalytics 360 accounts via `google_analytics_bq()`
 
 # 0.1.1